backend/requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
astunparse==1.6.3
async-timeout==5.0.1
asyncio==3.4.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cssselect==1.3.0
dnspython==2.7.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.7.0
gast==0.6.0
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.73.0
h5py==3.14.0
hf-xet==1.1.5
httplib2==0.22.0
huggingface-hub==0.33.4
idna==3.10
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.1
keras==3.10.0
libclang==18.1.1
lxml==5.4.0
lxml_html_clean==0.4.2
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.6.3
namex==0.1.0
networkx==3.4.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.26.4
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.3.1
pillow==11.2.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-whois==0.9.5
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-file==2.1.0
rich==14.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
sentence-transformers==5.0.0
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.7
sympy==1.14.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tf_keras==2.16.0
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.21.2
torch==2.2.2
tqdm==4.67.1
transformers==4.53.2
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.1
